Most actions are now bound.
There are minor cosmetic bugs that need to be taken care of:
KNOWN BUG:
after deleting the last entry in a list, auto-selection does not always work. (Sometimes work, though, I do not know why.)
this is particularly relevant for the "clean" action, which depends on selecting whatever remains and deleting it.
Using paths (inst of iter) and refresh could solve this bug.

I believe all the rest now works more or less.




git-svn-id: http://elogviewer.svn.sourceforge.net/svnroot/elogviewer@15 988d42b3-a221-0410-8e28-bfe4af7e9eeb