Simpler Elog/ElogItem/ElogRowItem API
API changes
-----------

Make Elog a simple immutable data structure.  It does not provide file
objects anymore and does not know the read/important state.

Only one Elog instance is stored per row in the row header ElogRowItem.

ElogItem fetches the appropriate data for the current column in
ElogRowItem.

read/important flag
-------------------

The model stores the current state of the read/important flags.  This
further allows to make the current row "unread" again.  State is loaded
on start in the `populate()` method and saved on quit in the
`saveSettings()` method.